requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
psycopg-binary==3.1.18
pytz==2024.1
sqlparse==0.4.4
